Add filter by trace level behave like logcat
- TraceLevel: new value "ALL". This value will be used to show all the trace
types. It is the default used on LynxConfig.
- Moved Trace#getTraceLevel to TraceLevel as it belongs to the Enum
- LynxView: add a spinner to show all the TraceLevel values. The style is
pretty basic, but it can be improved in following commits.
- Lynx: trace matches now uses TraceLevel too
- LynxPresenter: new update filter method for TraceLevel
- LynxConfig: has two new public methods
setFilterTraceLevel/getFilterTraceLevel. The reason to create another one
and not create a TraceFilter class it's because it will be a breaking change
on the public API.
I enforced to not pass a null value on the setters for the filter.